feat: 启用 denoising AE 训练模式并修复 retract_hpd 特征值类型问题
- train.py: 切换为 denoising AE，输入随机压缩特征值加噪
- train.py: 更新模型保存路径为 autoencoder.model
- util.py: retract_hpd 中特征值显式转为 torch.float64 以支持 clamp

+++ util.py
@@ -997,7 +997,7 @@ def retract_hpd(G, dLdlogG, t):
     #   Ω = V^H @ (dL/dlogG) @ V
 
     g, V = torch.linalg.eigh(G)  # g: (dim,) real ascending; V: (dim, dim) unitary
-    g = g.real.to(dtype)  # eigh 返回实特征值
+    g = g.real.to(torch.float64)  # eigh 返回实特征值, 保持实数类型以支持 clamp
 
     # Ω = V^H @ dLdlogG @ V
     Vh = V.conj().T
